[refactor]: Transformed webpack-server to normal server. Added necessary middleware to distinguish development and production environment
