+++ server/middleware/webpack-hmr.js
@@ -1,0 +1,9 @@
+import WebpackHotMiddleware from 'webpack-hot-middleware';
+
+const debug = require('debug')('kit:server:webpack-hmr');
+
+export default function ({ compiler }) {
+  debug('Enable Webpack Hot Module Replacement (HMR).');
+
+  return WebpackHotMiddleware(compiler);
+}
